src/agentry/agents/langgraph_studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langfuse
-e git+https://github.com/penlight-ai/agentry.git@main#egg=agentry
grandalf
python-dotenv